move fake agents physical interfaces to lowlevel/miinecraft/pyworld
